+++ server/controllers/dbController.js
@@ -1,12 +1,14 @@
 const { FirstName, LastName, Country } = require('../models');
-const dbController = [];
+// const dbController = [];
 
-const makeArray = (req, res, next) => {
+const dbController = {};
+
+dbController.makeArray = (req, res, next) => {
   res.locals.data = [];
   return next();
 };
   
-const getFirstNames = (req, res, next) => {
+dbController.getFirstNames = (req, res, next) => {
   const { firstName, fullName, fullNameMiddle, quantity } = req.query;
   if (!firstName && !fullName && !fullNameMiddle) return next();
   FirstName.aggregate([
@@ -31,9 +33,10 @@ const getFirstNames = (req, res, next) => {
      }
      return next(newErr)
     })
-};
+}
 
-const getMiddleNames = (req, res, next) => {
+
+dbController.getMiddleNames = (req, res, next) => {
   const {fullNameMiddle, quantity } = req.query;
   if (!fullNameMiddle) return next();
   FirstName.aggregate([
@@ -56,7 +59,7 @@ const getMiddleNames = (req, res, next) => {
   })
 };
 
-const getLastNames = (req, res, next) => {
+dbController.getLastNames = (req, res, next) => {
   const { lastName, fullName, fullNameMiddle, quantity } = req.query;
   if (!lastName && !fullName && !fullNameMiddle) return next();
   LastName.aggregate([
@@ -89,81 +92,7 @@ const getLastNames = (req, res, next) => {
   })
 };
 
-const getEmails = (req, res, next) => {
-  try{
-    const { email, quantity } = req.query;
-    if (!email) return next();
-    const tempArr = [...res.locals.data];
-    for (let i = 0; i < quantity; i++) {
-      let emailString = '';
-      const emailLength = Math.floor(Math.random() * 31 + 5);
-      
-      for (let i = 0; i < emailLength; i++) {
-        emailString += String.fromCharCode(Math.floor(Math.random() * 123 + 48));
-      }
-      
-      emailString = emailString.replace(/[^0-9A-Za-z]/g, '');
-      emailString += '@yeticrabs.com';
-      
-      if (tempArr[i]) {
-        tempArr[i].email = emailString;
-      } else {
-        const newObj = {
-          email: emailString
-        }
-        tempArr.push(newObj);
-      }
-    }
-    res.locals.data = tempArr;
-    return next();
-  }
-  catch {((err) => {
-    const newErr = {
-        log: 'error in getEmails',
-        message: { err: 'problem getting emails at this time'}
-    }
-    return next(newErr);
-  })
-  }
-};
-
-const getPhoneNumbers = (req, res, next) => {
-  try {
-    const { phoneNumber, quantity } = req.query;
-    if (!phoneNumber) return next();
-    const tempArr = [...res.locals.data];
-    
-    for (let i = 0; i < quantity; i++) {
-      let phoneNumString = '';
-      for (let i = 0; i < 10; i++) {
-        if (i===0) phoneNumString += '(';
-        phoneNumString += Math.floor(Math.random() * 10);
-        if (i===2) phoneNumString += ') ';
-        if (i===5) phoneNumString += '-';
-      }
-      if (tempArr[i]) {
-        tempArr[i].phoneNumber = phoneNumString
-      } else {
-        const newObj = {
-          phoneNumber: phoneNumString
-        }
-        tempArr.push(newObj);
-      }
-    }
-    res.locals.data = tempArr;
-    return next();
-  }
-  catch {((err) => {
-    const newErr = {
-        log: 'error in getPhoneNumbers',
-        message: { err: 'problem getting phone numbers at this time'}
-    }
-    return next(newErr);
-  })
-  }
-  };
-
-const getCountry = (req, res, next) => {
+dbController.getCountry = (req, res, next) => {
   const { country, quantity } = req.query;
   if (!country) return next();
   Country.aggregate([
@@ -196,11 +125,4 @@ const getCountry = (req, res, next) => {
   })
 }
 
-dbController.push(makeArray);
-dbController.push(getFirstNames);
-dbController.push(getMiddleNames);
-dbController.push(getLastNames);
-dbController.push(getEmails);
-dbController.push(getPhoneNumbers);
-dbController.push(getCountry);
 module.exports = dbController;
