+++ server/routers/apiRouter.js
@@ -1,10 +1,12 @@
 const express = require('express');
 const router = express.Router(); 
-const dbController = require('../controllers/dbController')
-
+const {makeArray, getFirstNames, getMiddleNames, getLastNames, getCountry} = require('../controllers/dbController')
+const {getEmails, getPhoneNumbers} = require('../controllers/otherDataController')
 // dbController is an array of all of our controller functions, NOT an object with controller methods
 
-router.get('/', dbController, (req, res) => {
+const controllers = [makeArray, getFirstNames, getMiddleNames, getLastNames, getCountry, getEmails, getPhoneNumbers]
+
+router.get('/', controllers, (req, res) => {
   return res.status(200).json(res.locals.data)
 });
